feat: Move FakeRequestsResponse to core.tests.base
