+++ marketplace/core/tests/__init__.py
@@ -1,0 +1,1 @@
+from .base import FakeRequestsResponse  # noqa
