+++ marketplace/core/types/channels/whatsapp/tests/test_views.py
@@ -6,22 +6,10 @@ from django.utils.crypto import get_random_string
 from django.test import override_settings
 from rest_framework import status
 
-from marketplace.core.tests.base import APIBaseTestCase
-from ..views import WhatsAppViewSet
+from marketplace.core.tests.base import APIBaseTestCase, FakeRequestsResponse
 from marketplace.applications.models import App
 from marketplace.accounts.models import ProjectAuthorization
-
-
-class FakeResponse(object):
-    def __init__(self, data: dict, error_message: str = None):
-        self._data = data
-        self.error_message = error_message
-
-    def raise_for_status(self):
-        pass
-
-    def json(self):
-        return self._data
+from ..views import WhatsAppViewSet
 
 
 class RetrieveWhatsAppTestCase(APIBaseTestCase):
@@ -111,7 +99,7 @@ class SharedWabasWhatsAppTestCase(APIBaseTestCase):
         self.set_responses()
 
     def set_responses(self):
-        self.debug_token_response = FakeResponse(
+        self.debug_token_response = FakeRequestsResponse(
             {
                 "data": {
                     "granular_scopes": [
@@ -125,7 +113,7 @@ class SharedWabasWhatsAppTestCase(APIBaseTestCase):
             }
         )
 
-        self.debug_token_without_business_response = FakeResponse(
+        self.debug_token_without_business_response = FakeRequestsResponse(
             {
                 "data": {
                     "granular_scopes": [
@@ -136,11 +124,13 @@ class SharedWabasWhatsAppTestCase(APIBaseTestCase):
         )
 
         error_message = "The access token could not be decrypted"
-        self.debug_token_error_response = FakeResponse({"data": {"error": {"message": error_message}}}, error_message)
+        self.debug_token_error_response = FakeRequestsResponse(
+            {"data": {"error": {"message": error_message}}}, error_message
+        )
 
         self.target_responses = [
-            FakeResponse(dict(id="1075799863265884", name="Fake target 1")),
-            FakeResponse(dict(id="1072999863265884", name="Fake target 2")),
+            FakeRequestsResponse(dict(id="1075799863265884", name="Fake target 1")),
+            FakeRequestsResponse(dict(id="1072999863265884", name="Fake target 2")),
         ]
 
     @property
